+++ monitor-services/bgpstreamlivetap/core/bgpstreamlive.py
@@ -22,6 +22,7 @@ from artemis_utils.envvars import REDIS_PORT
 from artemis_utils.envvars import REST_PORT
 from artemis_utils.rabbitmq import create_exchange
 from artemis_utils.redis import ping_redis
+from artemis_utils.redis import RedisExpiryChecker
 from artemis_utils.updates import key_generator
 from artemis_utils.updates import MformatValidator
 from artemis_utils.updates import normalize_msg_path
@@ -49,8 +50,6 @@ redis = redis.Redis(host=REDIS_HOST, port=REDIS_PORT)
 SERVICE_NAME = "bgpstreamlivetap"
 LIVE_PROJECT_MAPPINGS = {"ris": "ris-live", "routeviews": "routeviews-stream"}
 DEPRECATED_PROJECTS = ["caida"]
-
-# TODO: introduce redis-based restart logic (if no data is received within certain time frame)
 
 
 def start_data_worker(shared_memory_manager_dict):
@@ -514,6 +513,17 @@ def main():
             )
     except Exception:
         log.info("could not get configuration upon startup, will get via POST later")
+
+    # initiate redis checker
+    log.info("setting up redis expiry checker process...")
+    redis_checker = RedisExpiryChecker(
+        redis=redis,
+        shared_memory_manager_dict=bgpStreamLiveTapService.shared_memory_manager_dict,
+        monitor="bgpstreamlive",
+        stop_data_worker_fun=stop_data_worker,
+    )
+    mp.Process(target=redis_checker.run).start()
+    log.info("redis expiry checker set up")
 
     # start REST within main process
     bgpStreamLiveTapService.start_rest_app()