+++ docs/changelog.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Action and necessary logic to un-mitigate a hijack
 - Ability to reload targeted services based on what conf section changed
 - GET config endpoints to backend and monitor services
+- Redis no-data-received key expiry logic (default: 60 seconds)
 
 ### Changed
 - Decoupled microservice architecture for backend and frontend
@@ -24,6 +25,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Upgraded to new bgpstream containers (using also ris and routeviews live sources)
 - Full refactoring and upgrade of artemis-utils, refactored all imports (1.0.7)
 - Upgraded Flask-JWT-Extended to 3.25.0 to address JWT issue with new browsers
+- Upgraded artemis-utils to 1.0.8
 
 ### Fixed
 - Discard messages with timestamps in future datetime (updated to artemis-utils==1.0.5)